cleans up code and renames audio_generator
