+++ src/server/generate_audio/transcript.py
@@ -7,26 +7,26 @@ sample_transcript_list = [
         "speaker": "male_speaker_1",
         "text": "Hello everyone, I'm Tom Wilson. We're joined today by Maria Gonzalez, founder of three successful startups and author of 'Digital Disruption'."
     },
-    {
-        "speaker": "female_speaker_2",
-        "text": "Thanks for having me, Jennifer and Tom. I'm excited to share some insights about building businesses in today's rapidly changing landscape."
-    },
-    {
-        "speaker": "female_speaker_1", 
-        "text": "Maria, you started your first company right out of college. What inspired you to take that entrepreneurial leap so early?"
-    },
-    {
-        "speaker": "male_speaker_2",
-        "text": "I saw a gap in the market for mobile solutions for small businesses. Instead of waiting for someone else to fill it, I decided to take action. The worst that could happen was failure, and I figured I was young enough to recover from that."
-    },
-    {
-        "speaker": "male_speaker_1",
-        "text": "That's a great attitude. Can you tell us about the biggest challenges you faced in those early days?"
-    },
-    {
-        "speaker": "male_speaker_2",
-        "text": "Funding was definitely the biggest hurdle. Traditional investors weren't convinced that mobile-first solutions would take off. I had to bootstrap the company for the first two years while working part-time to pay the bills."
-    }
+    # {
+    #     "speaker": "female_speaker_2",
+    #     "text": "Thanks for having me, Jennifer and Tom. I'm excited to share some insights about building businesses in today's rapidly changing landscape."
+    # },
+    # {
+    #     "speaker": "female_speaker_1", 
+    #     "text": "Maria, you started your first company right out of college. What inspired you to take that entrepreneurial leap so early?"
+    # },
+    # {
+    #     "speaker": "male_speaker_2",
+    #     "text": "I saw a gap in the market for mobile solutions for small businesses. Instead of waiting for someone else to fill it, I decided to take action. The worst that could happen was failure, and I figured I was young enough to recover from that."
+    # },
+    # {
+    #     "speaker": "male_speaker_1",
+    #     "text": "That's a great attitude. Can you tell us about the biggest challenges you faced in those early days?"
+    # },
+    # {
+    #     "speaker": "male_speaker_2",
+    #     "text": "Funding was definitely the biggest hurdle. Traditional investors weren't convinced that mobile-first solutions would take off. I had to bootstrap the company for the first two years while working part-time to pay the bills."
+    # }
 ]
